Some fixes to the import scripts.
The scripts now print to stderr so that you can see their
progress when using dodo.py, like with mallet.

I also fixed a bug in creating the markup files, and now
the markup files are created in a hierarchical structure,
which was a problem before when there were 250,000 files in
the same directory.