Starting to fix a bug with the markup files.
I realized that I did the markup files wrong from the beginning.
Currently if you try to have more than one analysis for a given
dataset, things break because the database only stores one markup
file.  This commit fixes code that uses the markup file so that
it is tied to the analysis instead of just the document.  Database
changes and final fixes are next.