The main topic view now has proper tabs
I used jQuery UI to do it, which made the code incredibly simple.  The
theme can be changed with ThemeRoller, if that's desired.  This is just
the main part of the topic page, so there is still a lot more to change
if we want the whole UI to be consistently formatted like this.