Documents tab now mimics the topics tab
There are some rough edges, I suppose, but they are the same rough edges
as exist in the topics tab.  There are some places where there is only
one lower-level tab, so it's a bit awkward, but that's how it is.  At
least the code and the UI are consistent between the two tabs.
Attributes tab and words tab still to go.